Clean up file names, and add loats of documentation.
