+++ examples/java/build.gradle.kts
@@ -12,7 +12,7 @@ repositories {
 }
 
 dependencies {
-    implementation("org.audux.bgg:bggclient:0.5.1")
+    implementation("org.audux.bgg:bggclient:0.6.0")
 
     testImplementation(platform("org.junit:junit-bom:5.9.1"))
     testImplementation("org.junit.jupiter:junit-jupiter")
