+++ gradle.properties
@@ -3,4 +3,3 @@ org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\=
 
 #Kotlin
 kotlin.code.style=official
-
