+++ src/main/kotlin/org/audux/bgg/request/HotList.kt
@@ -16,7 +16,7 @@ package org.audux.bgg.request
 import io.ktor.client.request.get
 import io.ktor.client.statement.bodyAsText
 import io.ktor.http.appendPathSegments
-import org.audux.bgg.BggClient.InternalBggClient
+import org.audux.bgg.InternalBggClient
 import org.audux.bgg.common.Constants.PARAM_TYPE
 import org.audux.bgg.common.Constants.PATH_HOT
 import org.audux.bgg.common.Constants.XML2_API_URL
@@ -24,13 +24,8 @@ import org.audux.bgg.common.HotListType
 import org.audux.bgg.response.HotList
 import org.audux.bgg.response.Response
 
-/**
- * Hotness endpoint that retrieve the list of most 50 active items on the site filtered by type.
- *
- * @param type Single [HotListType] returning only items of the specified type, defaults to
- *   [HotListType.BOARD_GAME].
- */
-internal fun InternalBggClient.hotItems(type: HotListType?) = request {
+/** @see org.audux.bgg.BggClient.hotList */
+internal fun InternalBggClient.hotList(type: HotListType?) = request {
     client()
         .get(XML2_API_URL) {
             url {
