+++ src/main/kotlin/org/audux/bgg/response/HotList.kt
